backend/v1/requirements.txt
accelerate==0.25.0
aiofiles==23.2.1
aiohttp==3.9.1
aioredis==1.3.1
aiosignal==1.3.1
aiostream==0.5.2
altair==5.2.0
amqp==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asgiref==3.7.2
astroid==2.6.6
async-timeout==4.0.3
atomicwrites==1.4.1
attrs==23.1.0
Authlib==1.3.0
autobahn==23.6.2
Automat==22.10.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.1
beautifulsoup4==4.12.2
billiard==4.2.0
bitsandbytes==0.41.3.post2
blinker==1.7.0
blis==0.7.11
bs4==0.0.1
build==1.0.3
CacheControl==0.13.1
cachetools==5.3.2
catalogue==2.0.10
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==3.0.5
channels-redis==3.2.0
chardet==3.0.4
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.21
cleo==2.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.4
constantly==23.10.4
contextlib2==21.6.0
contourpy==1.2.0
crashtest==0.4.1
cryptography==41.0.7
cycler==0.12.1
cymem==2.0.8
daphne==3.0.2
dataclasses-json==0.5.8
datasets==2.16.0
de-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.7.0/de_core_news_sm-3.7.0-py3-none-any.whl#sha256=d88c737eb7eb766f730f6a2dcb99dfcdb81623e1e0d89a9c638a2182ac19c52e
decorator==5.1.1
deep-translator==1.11.4
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.7
distlib==0.3.8
distro==1.8.0
docarray==0.40.0
document==1.0
docx==0.2.4
docx2txt==0.8
dulwich==0.21.7
effdet==0.4.1
elastic-transport==8.11.0
elasticsearch==8.11.1
embedchain==0.1.43
emoji==2.9.0
faiss-cpu==1.7.4
fastapi==0.104.1
fastjsonschema==2.19.0
ffmpy==0.3.1
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
fleet-context==1.1.9
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.10.0
google-auth==2.25.2
googleapis-common-protos==1.62.0
googlesearch-python==1.2.3
googletrans==3.1.0a0
greenlet==3.0.1
grpcio==1.60.0
h11==0.9
h2==3.2.0
hiredis==2.3.2
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.*
httptools==0.6.1
httpx==0.13.3
huggingface-hub==0.19.4
humanfriendly==10.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==6.11.0
importlib-resources==6.1.1
incremental==22.10.0
iniconfig==2.0.0
installer==0.7.0
iopath==0.1.10
ipython==7.19.0
isort==5.13.2
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
keyring==24.3.0
kiwisolver==1.4.5
kombu==5.3.4
kubernetes==28.1.0
lancedb==0.4.0
langchain==0.0.336
langchain-community==0.0.8
langchain-core==0.1.5
langcodes==3.3.0
langdetect==1.0.9
langsmith==0.0.77
layoutparser==0.3.4
lazy-object-proxy==1.10.0
libretranslatepy==2.1.1
llama-index==0.8.69.post2
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
mccabe==0.6.1
mdurl==0.1.2
mmh3==4.0.1
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==1.6.1
opencv-python==4.8.1.78
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-asgi==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
orjson==3.9.10
overrides==7.4.0
packaging==23.2
pandas==2.1.4
parso==0.8.3
pathtools==0.1.2
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.3
pexpect==4.9.0
phonenumbers==8.13.27
pickleshare==0.7.5
pikepdf==8.10.1
Pillow==10.1.0
pkginfo==1.9.6
platformdirs==3.11.0
pluggy==1.3.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
portalocker==2.8.2
posthog==3.1.0
preshed==3.0.9
presidio-analyzer==2.2.351
presidio-anonymizer==2.2.351
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.7
ptyprocess==0.7.0
pulsar-client==3.3.0
py==1.11.0
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.19.1
pydantic==1.10.13
pydantic_core==2.14.5
pydub==0.25.1
Pygments==2.17.2
pylance==0.9.0
pylint==2.7.4
pyOpenSSL==23.3.0
pyparsing==3.1.1
pypdf==3.17.4
PyPDF2==3.0.1
pypdfium2==4.25.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pytesseract==0.3.10
pytest==6.2.5
pytest-django==4.1.0
pytest-mock==3.3.1
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-iso639==2023.12.11
python-magic==0.4.27
python-multipart==0.0.6
PyTranslation==1.0.3
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
ragclip==0.0.4
rank-bm25==0.2.2
rapidfuzz==3.5.2
ratelimiter==1.2.0.post0
redis==3.5.3
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
retry==0.9.2
rfc3986==1.5.0
rich==13.7.0
rpds-py==0.13.2
rsa==4.9
safetensors==0.4.1
schema==0.7.5
scikit-learn==1.3.2
scipy==1.11.4
semantic-version==2.10.0
semver==3.0.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
service-identity==23.1.0
shellingham==1.5.4
shortuuid==1.0.11
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.23
sqlparse==0.4.4
srsly==2.4.8
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
thinc==8.2.2
threadpoolctl==3.2.0
tiktoken==0.4.0
timm==0.9.12
tldextract==5.1.1
tokenizers==0.15.0
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
traitlets==5.14.0
transformers==4.36.2
translate==3.6.1
translation==1.0.5
translator==0.0.9
trove-classifiers==2023.11.29
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typer==0.9.0
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
ucall==0.5.1
Unidecode==1.3.7
unstructured==0.11.6
unstructured-client==0.15.1
unstructured-inference==0.7.18
unstructured.pytesseract==0.3.12
urllib3==1.26.18
uvicorn==0.24.0.post1
validators==0.21.0
vine==5.1.0
virtualenv==20.25.0
vvm==0.2.0
wasabi==1.1.2
watchdog==0.10.4
watchfiles==0.21.0
wcwidth==0.2.12
weasel==0.3.4
weaviate-client==3.17.1
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.1
wrapt==1.12.0
xxhash==3.4.1
yarl==1.9.3
zipp==3.17.0
zope.interface==6.1
